LayoutTests/plugins/netscape-plugin-wildcard-plugin.html

<html>
<head>
<script>
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.setWindowIsKey(false);
    }
    
</script>
<body>
<embed id="testPlugin" type="application/x-webkit-test-netscape-wildcard-plugin" width="200" height="200"></embed>
<div>
    This tests that a plugin which supports the wildcard mime type * is loaded correctly.
</div>
</body>
</html>
